+++ rust/src/stream/actors/user_command_persistence_actor.rs
@@ -17,16 +17,19 @@ pub struct UserCommandPersistenceActor {
 }
 
 impl UserCommandPersistenceActor {
-    pub async fn new(shared_publisher: Arc<SharedPublisher>) -> Self {
+    pub async fn new(shared_publisher: Arc<SharedPublisher>, preserve_existing_data: bool) -> Self {
         if let Ok((client, connection)) = tokio_postgres::connect(POSTGRES_CONNECTION_STRING, NoTls).await {
             tokio::spawn(async move {
                 if let Err(e) = connection.await {
                     eprintln!("connection error: {}", e);
                 }
             });
-            if let Err(e) = client.execute("DROP TABLE IF EXISTS user_commands;", &[]).await {
-                println!("Unable to drop user_commands table {:?}", e);
+            if !preserve_existing_data {
+                if let Err(e) = client.execute("DROP TABLE IF EXISTS user_commands;", &[]).await {
+                    println!("Unable to drop user_commands table {:?}", e);
+                }
             }
+
             if let Err(e) = client
                 .execute(
                     "CREATE TABLE IF NOT EXISTS user_commands (
